Thermo-nuclear: extract ollama cookies, codex pricing, simplify accent color
