+++ apps/desktop-tauri/src-tauri/src/main.rs
@@ -211,9 +211,6 @@ fn main() {
             commands::get_provider_region_options,
             commands::set_provider_workspace_id,
             commands::set_provider_gateway_url,
-            commands::set_provider_accent_color,
-            commands::get_provider_accent_color,
-            commands::get_provider_effective_accent_color,
             commands::get_provider_workspace_id,
             commands::get_gemini_cli_signed_in,
             commands::get_vertexai_status,
