+++ rust/src/core/mod.rs
@@ -2,6 +2,7 @@
 
 mod adaptive_refresh;
 mod aws_signing;
+mod codex_routed_pricing;
 mod cost_cache_budget;
 mod cost_pricing;
 pub mod curl_capture;
